pyproject.toml
[project]
name = "filoz-time-tracking-utils"
version = "0.1.0"
description = "Convert Timing App Excel exports into the biglep time tracking Google Sheet."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openpyxl>=3.1.0",
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    "python-dotenv>=1.0.0",
]
